// src/app/modules/admin/admin.interface.ts
import { Role, UserStatus } from "../../../generated/prisma/enums";

export interface IUpdateAdminPayload {
  admin: {
    name?: string;
    profilePhoto?: string;
    contactNumber?: string;
  };
}
export interface IChangeUserStatusPayload {
  userId: string;
  userStatus: UserStatus;
}

export interface iChangeUserRolePayload {
  userId: string;
  role: Role;
}